restaurant-service/requirements.txt
fastapi
uvicorn
asyncpg
redis
azure-servicebus
cryptography
python-dotenv
psycopg2-binary